Add Camera Shake
Added a camera shake effect on death when the characters slam to the ground. Mostly to mask the fact that I fixed the bug where the character stays in the air after dying in the air by having the character slam back to the ground position. Didn't have an animation for falling while dead, so yeah